+++ strato/core/strato-sequencer/exec-src/Main.hs
@@ -15,17 +15,13 @@ import Blockchain.Sequencer
 import Blockchain.Sequencer.CablePackage
 import Blockchain.Sequencer.Gregor
 import Blockchain.Sequencer.Monad
-import Blockchain.Strato.Model.ChainMember
 import Blockchain.Strato.Model.Options (flags_network)
 import Control.Concurrent (threadDelay)
 import Control.Concurrent.Async as Async
 import Control.Concurrent.STM
 import Control.Concurrent.STM.TMChan
 import Control.Monad
-import qualified Data.Aeson as Ae
-import Data.ByteString.Base64
 import qualified Data.ByteString.Char8 as C8
-import Data.Either.Extra
 import Flags
 import HFlags
 import Network.HTTP.Client (defaultManagerSettings, newManager)
@@ -34,6 +30,8 @@ import Network.Wai.Handler.Warp
 import Network.Wai.Middleware.Prometheus
 import Safe
 import Servant.Client
+import qualified Strato.Strato23.API as VC
+import qualified Strato.Strato23.Client as VC
 
 waitOnVault :: (Show a) => IO (Either a b) -> IO b
 waitOnVault action = do
@@ -57,7 +55,6 @@ main = do
   putStrLn $ "strato-sequencer validators: " ++ show validators
   putStrLn $ "strato-sequencer vault-proxy URL: " ++ show flags_vaultWrapperUrl
   putStrLn $ "strato-sequencer validatorBehavior: " ++ show flags_validatorBehavior
-  putStrLn $ "strato-sequencer certInfo: " ++ show flags_certInfo
 
   pkg <- atomically newCablePackage
   let kafkaClientId' = KP.KString $ C8.pack flags_kafkaclientid
@@ -81,21 +78,16 @@ main = do
   vaultWrapperUrl <- parseBaseUrl flags_vaultWrapperUrl
   let clientEnv = mkClientEnv mgr vaultWrapperUrl
 
-  --  Allow these flags to accept base64-encoded JSONs optionally
-  let b64decode inp = if isBase64 inp then (fromRight inp . decodeBase64) inp else inp
-      eSelf = (Ae.eitherDecodeStrict . b64decode) (C8.pack flags_certInfo) :: Either String ChainMemberParsedSet
-      !self = fromRight (error "invalid self cert info") eSelf
+  selfAddress <- do
+    addrAndKey <- waitOnVault $ runClientM (VC.getKey Nothing Nothing) clientEnv
+    return $ VC.unAddress addrAndKey
+
+  putStrLn $ "strato-sequencer nodeAddress: " ++ show selfAddress
 
   mCtx <-
     if not flags_blockstanbul
       then return Nothing
       else do
-        unless (self `elem` validators) . putStrLn $
-          "WARNING: NODEKEY does not correspond to a validator identity.\
-          \ This probably means that you are connecting to an existing network,\
-          \ and you are not one of the original validators of that network.\
-          \ If this is the case, please disregard this message. Otherwise,\
-          \ you may experience difficulty operating this node."
         unless (flags_blockstanbul_block_period_ms >= 0) . ioError . userError $
           "--blockstanbul_block_period_ms must be nonnegative"
         unless (flags_blockstanbul_round_period_s > 0) . ioError . userError $
@@ -106,7 +98,7 @@ main = do
         ckpt <- runGregorM gregorCfg $ initializeCheckpoint validators
         putStrLn $ "Checkpoint: " ++ show ckpt
 
-        return $ Just $ newContext ckpt self flags_validatorBehavior
+        return $ Just $ newContext ckpt (Just selfAddress) flags_validatorBehavior Nothing
 
   cht <- atomically newTMChan
 
@@ -125,6 +117,6 @@ main = do
             vaultClient = Just clientEnv
           }
   race_ (runTheGregor gregorCfg)
-    . race_ (runLoggingT (runSequencerM seqCfg mCtx sequencer))
+    . race_ (runLoggingT (runSequencerM seqCfg mCtx (sequencer validators)))
     . run flags_blockstanbul_port
     $ metricsApp